Set /dev/cpu_dma_latency value in hex
/dev/cpu_dma_latency accepts values in either s32 or hex format,
so convert 44 (dec) to 2c (hex).

I assume Google meant to set cpu dma latency to 44 in decimal
rather than hex because 44 is the lowest valid latency value for
C3 LPM state, according to qcom,latency-us in sdm845-pm.dtsi.

The pm qos request is only active when the fd is kept open, so
the dummy value of 100 does not really matter. Keep it as is.

Test: hexdump -C /dev/cpu_dma_latency

00000000  2c 00 00 00                                       |,...|
00000004

Test: kernel pr_err() in pm_qos_power_write() before
      pm_qos_update_request()

Inspired by Artem's commit "qos: Workaround for faulty libperfmgr
config".

[Hlcpereira]: Kona based platforms lowest value for cpu dma is 61, hence
61 -> 3d

Co-authored-by: Henrique Pereira <hlcpereira@pixelexperience.org>